requirements.txt
matplotlib>=3.5.0
numpy>=1.22.0
opencv-python>=4.5.5
pandas>=1.3.4
Pillow>=9.5.0
scipy>=1.8.1
torch>=1.12.1
torchvision>=0.13.1


